fix: GetFoodWorkers() should return uint32_t

+++ src/sc2api/sc2_client.cc
@@ -169,7 +169,7 @@ public:
     uint32_t GetFoodCap() const final { return food_cap_; }
     uint32_t GetFoodUsed() const final { return food_used_; }
     uint32_t GetFoodArmy() const final { return food_army_; }
-    uint32_t GetFoodWorkers() const final { return static_cast<float>(food_workers_); }
+    uint32_t GetFoodWorkers() const final { return food_workers_; }
     uint32_t GetIdleWorkerCount() const final { return idle_worker_count_; }
     uint32_t GetArmyCount() const final { return army_count_; }
     uint32_t GetWarpGateCount() const final { return warp_gate_count_; }
